+++ app/components/chat-list.tsx
@@ -1,5 +1,5 @@
 import DeleteIcon from "../icons/delete.svg";
-import styles from "./home.module.scss";
+import styles from "./chatHomepage.module.scss";
 import {
   DragDropContext,
   Droppable,
