--- app/components/employeeCard.module.scss
@@ -1,56 +1,0 @@
-$defaultWhite: #fff;
-$defaultBlack: #000;
-$secondary-4-default: #ede8d7;
-
-.list {
-  list-style-type: none;
-  justify-self: center;
-  padding: 0;
-}
-
-.employeeCard {
-  background-color: $defaultWhite;
-  margin: 0;
-  padding-inline: 2rem;
-  width: 13rem;
-  height: 10rem;
-
-  text-decoration: none;
-
-  border-radius: 1.1875rem;
-  border: 1px solid $secondary-4-default;
-  background: $defaultWhite;
-  box-shadow: 0px 4px 36px 0px rgba(0, 0, 0, 0.05);
-
-  display: flex;
-  gap: 2.25rem;
-  align-items: center;
-  justify-content: space-between;
-
-  transition: 0.3s all ease;
-
-  &:hover {
-    border-color: var(--primary);
-  }
-
-  & img {
-    width: 100px;
-    height: 100px;
-  }
-
-  & .employeeName {
-    color: $defaultBlack;
-    font-family: Recoleta;
-    font-size: 1.125rem;
-    font-style: normal;
-    font-weight: 400;
-  }
-
-  & p {
-    color: $defaultBlack;
-    font-family: Graphik;
-    font-size: 0.625rem;
-    font-style: normal;
-    font-weight: 300;
-  }
-}
