--- app/components/employeeCv.module.scss
@@ -1,3 +1,0 @@
-.RequirementInput{
-    min-height: 200px;
-}
