--- app/components/employees.module.scss
@@ -1,29 +1,0 @@
-.employee-buttons {
-  display: flex;
-}
-
-@mixin employee-list {
-  overflow-y: auto;
-  list-style: none;
-  padding-inline: 1.5rem;
-  display: grid;
-  align-items: center;
-  gap: 1.5rem;
-}
-
-.employee-list {
-  @include employee-list();
-  grid-template-columns: 1fr 1fr;
-}
-
-.employee-list-tight {
-  @include employee-list();
-  width: 100%;
-  grid-template-columns: repeat(auto-fill, minmax(16rem, 1fr));
-}
-
-@media (max-width: 800px) {
-  .employee-list {
-    grid-template-columns: 1fr;
-  }
-}
