+++ app/components/salesGPT.module.scss
@@ -1,0 +1,1 @@
+@import "./window.scss";
