+++ app/utils.ts
@@ -91,3 +91,7 @@ export function getCurrentVersion() {
 export function getEmojiUrl(unified: string, style: EmojiStyle) {
   return `https://cdn.staticfile.org/emoji-datasource-apple/14.0.0/img/${style}/64/${unified}.png`;
 }
+
+export function aliasFromEmail(email: string | undefined) {
+  return email?.split("@")[0];
+}
